test: move resolved_pkg_group_with_dummy_package to flox-rust-sdk
