+++ cli/flox/src/commands/init/mod.rs
@@ -763,8 +763,6 @@ async fn try_find_compatible_version(
 #[cfg(test)]
 mod tests {
 
-    use flox_rust_sdk::data::System;
-    use flox_rust_sdk::providers::catalog::{CatalogPage, ResolvedPackageGroup};
     use indoc::indoc;
     use pretty_assertions::assert_eq;
 
@@ -790,48 +788,6 @@ mod tests {
                     None
                 },
             }
-        }
-    }
-
-    // This function should really be a #[cfg(test)] method on ResolvedPackageGroup,
-    // but you can't import test features across crates
-    pub fn resolved_pkg_group_with_dummy_package(
-        group_name: &str,
-        system: &System,
-        install_id: &str,
-        pkg_path: &str,
-        version: &str,
-    ) -> ResolvedPackageGroup {
-        let pkg = PackageResolutionInfo {
-            attr_path: pkg_path.to_string(),
-            broken: Some(false),
-            derivation: String::new(),
-            description: None,
-            install_id: install_id.to_string(),
-            license: None,
-            locked_url: String::new(),
-            name: String::new(),
-            outputs: vec![],
-            outputs_to_install: None,
-            pname: String::new(),
-            rev: String::new(),
-            rev_count: 0,
-            rev_date: chrono::offset::Utc::now(),
-            scrape_date: chrono::offset::Utc::now(),
-            stabilities: None,
-            unfree: None,
-            version: version.to_string(),
-            system: system.parse().unwrap(),
-        };
-        let page = CatalogPage {
-            packages: Some(vec![pkg]),
-            page: 0,
-            url: String::new(),
-            complete: true,
-        };
-        ResolvedPackageGroup {
-            name: group_name.to_string(),
-            page: Some(page),
         }
     }
